Refactor(*): JPA를 Domain 모듈로 이동
- JPA 의존성을 Infrastructure 모듈에서 Domain 모듈로 이동
- Infrastructure 모듈에 있던 매핑관련 클래스들 모두 삭제
- Infrastructure 모듈에는 일단 클래스 파일이 없고 yml 설정파일만 존재
- Domain 모듈의 Entity를 POJO에서 JPA가 관리하는 Entity로 변경
- JpaRepository interface 를 api 모듈에서 그대로 사용한다면 에러가 발생해 repository를 구현한 구현체를 둠으로써 에러를 해결하고 영속성 관련 예외처리도 구현
