+++ jabiseo-domain/build.gradle
@@ -9,6 +9,6 @@ jar {
 }
 
 dependencies {
-
+    implementation 'org.springframework.boot:spring-boot-starter-data-jpa'
     implementation project(":jabiseo-common")
 }
