+++ jabiseo-domain/src/main/java/com/jabiseo/certificate/Certificate.java
@@ -1,21 +1,31 @@
 package com.jabiseo.certificate;
 
 
+import jakarta.persistence.Column;
+import jakarta.persistence.Entity;
+import jakarta.persistence.Id;
+import jakarta.persistence.Table;
+import lombok.AccessLevel;
 import lombok.Builder;
 import lombok.Getter;
+import lombok.NoArgsConstructor;
 
+@Entity
 @Getter
+@Table(name = "certificate")
+@NoArgsConstructor(access = AccessLevel.PROTECTED)
 public class Certificate {
 
-    private String certificateId;
+    @Id
+    @Column(name = "certificate_id")
+    private String id;
 
     private String name;
 
     @Builder
-    public Certificate(String certificateId, String name) {
-        this.certificateId = certificateId;
+    public Certificate(String id, String name) {
+        this.id = id;
         this.name = name;
     }
 
 }
-
