+++ jabiseo-domain/src/main/java/com/jabiseo/certificate/JpaCertificateRepository.java
@@ -1,0 +1,6 @@
+package com.jabiseo.certificate;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface JpaCertificateRepository extends JpaRepository<Certificate, String> {
+}
