+++ jabiseo-infrastructure/build.gradle
@@ -9,7 +9,6 @@ jar {
 
 dependencies {
 
-    implementation 'org.springframework.boot:spring-boot-starter-data-jpa'
     implementation 'org.springframework.boot:spring-boot-starter-data-redis'
 
     runtimeOnly 'com.mysql:mysql-connector-j'
