--- jabiseo-infrastructure/src/main/java/com/jabiseo/database/CertificateMapper.java
@@ -1,22 +1,0 @@
-package com.jabiseo.database;
-
-import com.jabiseo.certificate.Certificate;
-
-
-public class CertificateMapper {
-
-    static Certificate jpaToDomain(JpaCertificateEntity entity) {
-        return Certificate.builder()
-                .certificateId(entity.getId())
-                .name(entity.getName())
-                .build();
-    }
-
-    static JpaCertificateEntity domainToJpa(Certificate domain) {
-        return JpaCertificateEntity.builder()
-                .id(domain.getCertificateId())
-                .name(domain.getCertificateId())
-                .build();
-    }
-
-}
