--- jabiseo-infrastructure/src/main/java/com/jabiseo/database/JpaCertifcateRepository.java
@@ -1,7 +1,0 @@
-package com.jabiseo.database;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface JpaCertifcateRepository extends JpaRepository<JpaCertificateEntity, String> {
-
-}
